requirements.txt
streamlit==1.25.0
PyYAML==6.0.1
st-annotated-text==4.0.0